refactor: add filters to phd package
